backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pg8000
pydantic
pydantic[email]
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt
python-dotenv
email-validator
httpx
pytest
gunicorn
google-genai
